refactoring header
