+++ src/Components/header/index.jsx
@@ -3,6 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import { useUser } from '../../app/context';
 import { Bar } from '../bar/index';
 import styles from './index.module.css';
+import { Button } from 'antd'
 
 export const Header = () => {
     const navigate = useNavigate();
@@ -18,6 +19,17 @@ export const Header = () => {
         }
     }, [user, authorized]);
 
+    const handleExit = () => {
+        // сохраняем данные юзера в Local Storage
+        const currentUser = JSON.parse(localStorage.getItem('current-user'));
+        if (currentUser) {
+            const updatedUser = { ...currentUser, authorized: false };
+            localStorage.setItem('current-user', JSON.stringify(updatedUser));
+        }
+
+        navigate('/login', { replace: true });
+    };
+
     return (
         <div className={styles.header}>
             <button
@@ -32,6 +44,16 @@ export const Header = () => {
             </button>
             <Bar />
             <div className={styles.inform}>
+                <Button
+                    onClick={() => (authorized ? handleExit() : navigate('/login', { replace: false }))}
+                    type="primary"
+                    htmlType="submit"
+                    color="default"
+                    variant="solid"
+                    className={styles.logBtn}
+                >
+                    {authorized ? 'Exit' : 'Login / Registration'}
+                </Button>
                 <p className={styles.logInfo}>{displayUser}</p>
             </div>
         </div>
