+++ src/Components/header/index.module.css
@@ -18,7 +18,7 @@
 .inform {
     display: flex;
     align-items: center;
-    gap: 3px;
+    gap: 15px;
 }
 
 .loginInfo {
