+++ src/Pages/home/index.jsx
@@ -1,19 +1,17 @@
 import React from 'react';
 import styles from './index.module.css';
 import { useNavigate } from 'react-router-dom';
-import { useUser } from '../../app/context';
+// import { useUser } from '../../app/context';
 import { Loader } from '../../Components/loader';
 import { MovieCard } from '../../Components/movie-card';
 import Searchfild from '../../Components/serch-input';
 import { ErrorPage } from '../404/index'
-import { Button } from 'antd';
 
 import { useGetMoviesQuery } from '../../features/movieApiSlice';
 
 export const Home = () => {
     const navigate = useNavigate();
     const { data, error, isLoading } = useGetMoviesQuery();
-    const { authorized } = useUser();
 
     const handleSearch = (word) => {
         console.log('Searching for:', word);
@@ -28,34 +26,12 @@ export const Home = () => {
         console.log(data)
     };
 
-    const handleExit = () => {
-        // сохраняем данные юзера в Local Storage
-        const currentUser = JSON.parse(localStorage.getItem('current-user'));
-        if (currentUser) {
-            const updatedUser = { ...currentUser, authorized: false };
-            localStorage.setItem('current-user', JSON.stringify(updatedUser));
-        }
-
-        navigate('/login', { replace: true });
-    };
-
     return (
         <div className={styles.home}>
             {isLoading ? (
                 <Loader />
             ) : (
                 <div className={styles.movieList}>
-
-                    <Button
-                        onClick={() => (authorized ? handleExit() : navigate('/login', { replace: false }))}
-                        type="primary"
-                        htmlType="submit"
-                        color="default"
-                        variant="solid"
-                        className={styles.logBtn}
-                    >
-                        {authorized ? 'Exit' : 'Login / Registration'}
-                    </Button>
 
                     <h1>Movies</h1>
                     <Searchfild onSearch={handleSearch} />
